Revert "first commit"
This reverts commit 11ab7aab

+++ src/main/java/ir/redmind/paasho/web/rest/mock/EventResources.java
@@ -325,7 +325,7 @@ public class EventResources {
             event1.setDate(String.valueOf(e[4]));
             event1.setLatitude((Double) e[6]);
             event1.setLongitude((Double) e[7]);
-            event1.setView((Long) e[8]);
+            event1.setView(Long.valueOf((Integer) e[8]));
 
             eventDTOS.add(event1);
         });
